rtl8852au: Silence the spamming of the logs
This driver was spamming the logs with useless information, such that the
real log info was getting lost. To do this, the driver log levels were
reordered in increasing rank of importance. We now have the following:

/* driver log level*/
enum {
	_DRV_NONE_,
	_DRV_ERR_,
	_DRV_WARNING_,
	_DRV_INFO_,
	_DRV_DEBUG_,
	_DRV_ALWAYS_,
	_DRV_MAX_
};

By default, the level is set to 3 (_DRV_WARNING_). If more detail is wanted,
the Makefile will need to be modified. In the future, I will be adding a
module parameter, and removing a lot of them that are currently available,
but of little use in the field.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>